tests/wpt/tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.negativedest.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.drawImage.negativedest</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.negativedest</h1>
<p class="desc">Negative destination width/height represents the correct rectangle</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  const response = await fetch('/images/ggrr-256x256.png');
  const blob = await response.blob();
  const bitmap = await createImageBitmap(blob);
  ctx.drawImage(bitmap, 100, 78, 50, 50, 0, 50, 50, -50);
  ctx.drawImage(bitmap, 100, 128, 50, -50, 100, 50, -50, -50);
  _assertPixelApprox(canvas, 1,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 1,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 98,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 98,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 48,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 48,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 51,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 51,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 25,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 75,25, 0,255,0,255, 2);

}, "Negative destination width/height represents the correct rectangle");
</script>
